+++ modulesAndFiles/quickSort.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // quick sort
-int* quickSort(int* const array, const size_t firstLimitation, const size_t secondLimitation);
+int* quickSort(int* const array, const int firstLimitation, const int secondLimitation);
